Update Google font references and add 'yup' dependency
These changes include updating the references to Google fonts, with the Bungee Inline now being replaced by Roboto in both the main application and Storybook. Additionally, the 'yup' package has been added as a new dependency to the application. A new text input component using styled-components is created with corresponding Storybook stories, and the global styles setting font family is updated as well.